Sacado:  Add test and fix for bug #6135.
A sacado user reported a bug where use of any of the SLFad types could
cause incorrect derivative computations when the derivative array needed
to be resized and it appeare on both the left and right-hand sides of an
expression.  I added a test for this, demonstrated it failed, and then
fixed it.  I took the opportunity to simplify how derivative components
are zeroed out as well.  I also fixed a bunch of white space problems.

Build/Test Cases Summary
Enabled Packages: Sacado
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=500,notpassed=0 (4.19 min)
1) SERIAL_RELEASE => passed: passed=502,notpassed=0 (4.16 min)
2) MPI_DEBUG_SS => passed: passed=500,notpassed=0 (9.78 min)