STK: Disable stk_sddm in Trilinos build
There is an dependency in sddm to the Sierra parser that somehow as able to link
previously. This dependency prevents the Trilinos build from working and appears
to be core to sddm. For now I am disabling sddm in the Trilinos build. Since it
didn't exist in the old verison of STK that Trilinos has this shouldn't be much
of an issue. Once this is resolved sddm can be re-enabled.

(From Sierra)