MueLu: UC: fixed inefficiency in phase 1
The original version of adding nodes to the GRAPH ordering may have
added the neighbors of the rootCandidate even if it was accepted. The
problem was that addition was done in the same loop as changing the
status of those neighbor nodes to AGGREGATED.

I separated loops, and simplified the code by joining two cases
(accepted/non-accepted).