Anasazi: Fixed Doxygen for block Krylov-Schur example that uses Amesos.
In the example code itself, when Doxygen finds comments that use
Doxygen, it strips it out of the code listing.  Thus, the code listing
had a lot of useful comments missing.  I fixed this by "de-Doxygening"
the comments, and moving some comments and the Doxygen-formatted
description of the example to AnasaziBlockKrylovSchurSolMgr.hpp (where
the link for the example was).