MueLu: UC: separated Dirichlet and isolated points
From aggregation factory point of view, Dirichlet points are detected by
somebody else/provided by user. In the UncoupledAggregationFactory they
are marked as BOUNDARY (they are actually detected in the
CoalesceDropFactory). Depending on user's wish, we may keep or drop
them. The responsible algorithm is PreserveDirichlet algorithm. They are
marked as IGNORED, and if we decide to keep them, they are put into a
singleton aggregates.

IGNORED is a new type which says that any consecutive aggregation
algorithm should skip them from any consideration. The reason is that if
your graph is non-symmetric, some aggregation routine could have a
connection to the Dirichlet node, and put it into a different aggregate.
This leads to problems in all cases: if Dirichlet was aggregated, it
would be aggregated twice into different aggregates; if it was not
aggregated, it would have incorrect vertex2AggId (-1).

On the other hand, isolated points are detected in
IsolatedNode algorithm, and are universally excluded by setting them to
IGNORED.

The code is pretty fragile at this point. The order matters: Dirichlet
algorithm should be in the beginning, isolated at the end. I'm pretty
sure that there are multiple bugs hiding here for non-symmetric
problems. But for symmetric ones, things should be ok.