MueLu: warn when  aggregation stagnates
When doing no repartitioning, or trying to construct hierarchy for a
diagonal matrix with isolated nodes going into aggregates, the
coarsening ratio can get close to 1. In this case, no real progress is
made, and it is much more preferrable to stop.

It is not an ideal solution, though. It may be better to stop a level or
two earlier, but currently there is no mechanism in the Hierarchy to
roll back levels.