PyTrilinos: Add C++ support for the DistArray Protocol
The purpose of C++ support for the DistArray Protocol is to take a
DistArray protocol object, and applying the rules of the protocol,
convert it into a C++ object that contains the same information.  If
there are errors during the conversion, a PythonException will be
thrown.

This support includes a DistributionType enumeration, typedefs for
index types and padding types, a DimensionDictionary class and a
DistArrayProtocol class.