Tpetra: MultiVector: Attempt to fix Kokkos refactor dot() issue.
Tpetra tests pass, but some Ifpack2 and MueLu tests don't pass, with
the Kokkos refactor version of Tpetra.  Christian and I traced the
issue to dot(), which dot() doesn't currently work for the
single-vector case.  The problem appears to be with
Kokkos::parallel_reduce when the functor has a final() method that
writes to a device View.  This commit makes the method more
conservative about error checking, but doesn't fix the issue.