PyTrilinos: Made Teuchos Comm wrappers more useful
First I changed the default (and only) template argument from long to
int.  Other classes that use Teuchos::Comm use the int template
parameter, so this is the appropriate way to go.  Next I changed thee
internal storage method for these wrapped objects to Teuchos::RCP<
Teuxhos::Comm > because this is how they are always stored in Trilinos
packages.  This propagated to Teuchos::Describable and
Teuchos::LabeledObject, and that spread form there...