STK Add scripts to convert stk namespace to stk_classic
Adding the scripts that were used to help convert Trilinos packages that use the
old STK to the namespace now in use by STKClassic. This is mostly for
convenience to allow users who only have a copy of the code(such as a Trilinos
release tarball) easy access to these scripts.

(From Sierra)