Stokhos:  Store length-1 expansions in a local value in DynamicStorage.
This adds local storage of length-1 expansions in DynamicStorage by
storing the value in a local variable instead of a dynamically allocated
array.  This makes manipulating length-1 expansions, such as the
mean-based preconditioner, much faster and makes building a mean-based
preconditioner on CUDA feasible.  There is a serious downside to this
approach though for CUDA since the pointer for the coefficient array can
become out of sync with the value when copying PCE variables to the
device.  It just means you have to have an overload of every CUDA kernel
for PCE.