Fixed a memory leak in SaddleContainer.hpp
SaddleContainer and SaddleOperator are both used to solve the
saddle-point problem arising at each iteration of TraceMin.
SaddleContainer stores a distributed set of vectors (the
approximate eigenvectors) and a small dense matrix (the Lagrange
multipliers).  It was storing the small dense matrix as a
ScalarType**, which introduced a memory leak.  I modified the code
to use a SerialDenseMatrix instead, which caused a separate (but
less terrible) problem.  SerialDenseMatrix does not support
noncontiguous views, so SaddleContainer no longer supports them
either.  The MINRES residing in Anasazi has been slightly modified
to reflect this.  SaddleContainer can NOT be used with Belos due
to this change (but TraceMin has some funny compatability issues
with Belos to begin with.  It has to do with the Ritz shifts and
the fact that we're solving multiple linear systems with different
right hand sides and different operators as well.  This is also
the reason a separate implementation of MINRES is currently living
in Anasazi; this implementation is capable of handling different
operators for each RHS.)

Build/Test Cases Summary
Enabled Packages: Anasazi
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=249,notpassed=0 (25.35 min)
1) SERIAL_RELEASE => passed: passed=250,notpassed=0 (6.86 min)