Belos::MultiVecTraits: Force Tpetra specialization to use view semantics
The Tpetra specialization of MultiVecTraits now forces MultiVector
objects that Belos itself creates to use view semantics.  This change
makes life easier for Tpetra::Experimental::BlockCrsMatrix, which
wants to view the MultiVector arguments to apply() as
BlockMultiVector.  The latter prefers the MultiVector that it views to
have view semantics.  (My recent commit relaxes this requirement, but
it's still better to favor view semantics, since this is where Tpetra
is going anyway.)

The two overloads of CloneView and the two overloads of
CloneViewNonConst are the only exceptions to the changes in this
commit.  For some reason I currently don't understand, the Belos
MVOpTester and/or OrthoManager tests fail if I make the returned view
have view semantics.  This is true for both the "classic" and Kokkos
refactor versions of Tpetra.  I don't know why this is the case.
Belos shouldn't care whether Tpetra uses copy or view semantics, and
the Kokkos refactor version of Tpetra _always_ uses view semantics.
Nevertheless, the tests fail, so I will leave the following line
commented out for now.  This is true for both CloneView overloads as
well as both CloneViewNonConst overloads.

Thanks to Travis Fisher for discovering and reporting this issue!

Note: "Copy" vs. "view" semantics refer to the default behavior of
MultiVector's copy constructor and operator=.  See the documentation
of MultiVector::{get,set}CopyOrView for details.  It sounds confusing,
but a returned "view of a MultiVector" -- e.g., the result of
subView(NonConst) or offsetView(NonConst) -- may have either copy or
view semantics, in the "classic" version of Tpetra.  The Kokkos
refactor version of MultiVector _always_ has view semantics; it is not
allowed to have copy semantics.

Build/Test Cases Summary
Enabled Packages: KokkosCore, KokkosCompat, KokkosContainers, KokkosLinAlg, KokkosMpiComm, KokkosAlgorithms, Tpetra, Ifpack2, Belos, Amesos2, Zoltan2
0) MPI_DEBUG => passed: passed=153,notpassed=0 (4.14 min)
1) SERIAL_RELEASE => passed: passed=127,notpassed=0 (0.39 min)
2) MPI_DEBUG_KOKKOS_REFACTOR => passed: passed=215,notpassed=0 (6.08 min)
3) MPI_RELEASE_KOKKOS_REFACTOR => passed: passed=206,notpassed=0 (4.26 min)
Other local commits for this build/test group: efa5735, 5fae25c, a762b21