Tpetra: CrsGraph: Added reindexColumns method; currently not working
This commit affects both the "classic" and Kokkos refactor versions of
Tpetra.

This commit adds a new method to CrsGraph, called reindexColumns.
This method is supposed to apply a new column Map to the graph, and
reindex the graph's column indices in place.  The method currently
throws std::logic_error; I have not yet implemented it.  In the best
tradition of test-driven development, I have added a unit test which
exercises this method.  Thus, the unit test WILL FAIL; this commit
builds, but does not pass that unit test.  The unit test lives in
tpetra/test/CrsGraph/CrsGraph_ReindexColumns.cpp.  I have verified
that the unit test fails due only to the method not being implemented.

The unit test depends on CrsGraph::clone.  I found a minor bug in how
clone() accesses its input RCP<ParameterList>, and fixed it.  (It must
be that none of the unit tests exercised clone() with a null input
RCP<ParameterList>.)  Thus, this commit also contains a fix to
clone().