MueLu: Ifpack2: Amesos2: fix ETI and GO=longlong
This is a work-around to get ETI working in in MueLu with GO=long long,
just until a proper Trilinos ETI system is in place.

I've tested by configuring with two different cmake scripts with ETI and
compiling/running MueLu_TwoBillionTest.exe:

1)
  -D Trilinos_ENABLE_EXPLICIT_INSTANTIATION:BOOL=ON \
  -D Teuchos_ENABLE_LONG_LONG_INT:BOOL=ON \

2)
  -D Trilinos_ENABLE_EXPLICIT_INSTANTIATION:BOOL=ON \

Case 1 causes the TwoBillion test to be built, case 2 does not.