Trilinos / Kokkos: Adding nvcc_wrapper
nvcc_wrapper is a script which can be used as the CXX compiler for Trilinos.
It calls nvcc but treats the arguments correctly, that is it seperates arguments
into ones know by nvcc, linker arguments and unknown arguments which presumably
should be piped to the host compiler only. It also identifies all source files
and adds them to the compile line only after a -x cu. The goal of this is to
be able to compile *.cpp, *.cxx and so on files which contain Cuda code via
Kokkos (essentially every KOKKOS_INLINE_FUNCTION is Cuda code). Just using nvcc
directly doesn't work since for example configure of Trilinos did not parse. There
are two reasons for the latter: (i) nvcc does not understand all gcc arguments, they
need to be preceded by a -Xcompiler. (ii) all input files after an -x cu are treated
as cuda source files, this includes *.o and *.a files. So a configure test like:

nvcc -O3 -x cu main.cpp object.o -Wall

would fail because nvcc treats object.o as a source file and doesn't understand -Wall.
The wrapper script reworks this into:

nvcc -O3 object.o -Xcompiler -Wall -x cu main.cpp

which works. With this script it is now possible to compile all of Trilinos with nvcc
and have the CudaWrapperNode as the DefaultNodeType for the software stack build on top
of Tpetra.

Things you need to do:
(i) set nvcc_wrapper as your cxx and mpi_cxx compiler (but don't replace your "C" compiler)
  -D CMAKE_CXX_COMPILER:FILEPATH="nvcc_wrapper" \
  -D MPI_CXX_COMPILER:FILEPATH="nvcc_wrapper" \
(ii) add CXX flags to enable so that MPI is found and the CUDA architecture is set
  -D CMAKE_CXX_FLAGS:STRING="-I${MPI_PATH}/include -arch=sm_35" \