Allowing generation of export files for testing purposes (Trilinos #6168)
The generation of export files was turned off several years ago when it was
taking too long.  This has seen be fixed (don't remember bug number) so I am
enabling this again.

It turns out that several projects (e.g. Albany) are using export files from
Trilinos so it is important to test this before pushing.