Tpetra: VbrMatrix: Removed LocalMatOps template parameter.
This is technically a backwards-incompatible change.  The LocalMatOps
template parameter was intended to be an extension point for
introducing third-party implementations of sparse matrix-vector
multiply and other kernels.  However, I don't know of any users who
ever exercised anything other than the default value of this template
parameter.  It complicates downstream packages quite a bit, and we're
deprecating it in CrsGraph and CrsMatrix anyway.  Thus, I am removing
it from VbrMatrix.  If users complain, I can always reintroduce this
template parameter with a default value of void.  I prefer to leave it
out, though.  Its omission should also reduce compilation times (by
giving the compiler less to parse).

I also took the opportunity to remove some unnecessary uses of the
KokkosClassic namespace (and a class therein) in Tpetra's tests.