Kokkos: Cuda implementation change that kokkos uses the y-Dim of block for threads
This change effectivly swap threadIdx.y for threadIdx.x in the internal implementation
of Kokkos (and associated things such as blockDim.x and the cuda launch dispatch).
This will allow us to use the x dimension of the block for the vector level parallelism.

Note: if a code used threadId.x directly for a Cuda specialisation of a kernel that needs
to be changed.