Kokkos: Allow C++11 if cuda version is >= 6.5
CUDA 6.5 inofficially supports C++11 in device code. This includes
stuff like lambdas and auto types. It does not enable dispatch of
lambdas from the host, since lambdas are not allowed as arguments
to __global__ functions.

We introduce a new internal typedef: KOKKOS_HAVE_CXX11_DISPATCH_LAMBDA
which can be used to guard such lambda dispatches for example in tests
against be compiled for CUDA.