Ifpack2: Fixed DenseContainer unit test.
The test was depending on Tpetra::Vector's one-argument copy
constructor doing a deep copy.  This is no longer true when Tpetra is
using view semantics, as it always does with the Kokkos refactor
version of Tpetra.  I fixed the test and now it passes.  I think
DenseContainer was actually correct, but I went ahead and made some of
its implementation details a bit simpler (e.g., no need to create a
Vector by RCP if its scope is local to the method anyway).