MueLu: cosmetic update to SetParameterList
Call GetParameterList before setting the parameters. If one calls
SetParameterList before GetParameterList, that would mean that
paramList_ has not been initialized yet. Therefore, the parameter would
be put in it in the order user provided, and not in the order of valid
parameter list. We would like to have consistency in the order, so we do
this extra call, which is no-op if paramList_ has already been
initialized.