Ifpack2: Removed test that depended on Tpetra::HybridPlatform.
This will facilitate deprecation of the KokkosClassic subpackage.  The
test wasn't being built because it depended on Anasazi, which is no
longer an optional test dependency of Ifpack2.

(It might make sense for Ifpack2 itself, not just its test, to depend
on Anasazi, for example to compute spectral radius bounds.  However,
the opposite direction of dependency also makes sense, for example for
solvers with shift and invert.  The right way to break this circular
dependency would be through Stratimikos or something like it.)