Kokkos: usecase/examples fixed build issues when using the nvcc_wrapper
When building Kokkos code with nvcc, while KOKKOS_HAVE_CUDA is defined,
Kokkos_Cuda.hpp must be included. This was not the case for the feint
example where the Cuda build was handled in a special .cu file. When
building with nvcc (i.e. through the nvcc_wrapper) KOKKOS_INLINE_FUNCTION
is defined as __device__ __host__ hence the parser will see device code.
When Kokkos_Cuda.hpp is not included the verification of ActiveExecutionSpace
to HostSpace fails, because that is only defined in KOkkos_CudaSpace.hpp.

These functions then have runtime abords, but they must exist since they are
parsed. Generally the issue can be resolved by always including Kokkos_Core.hpp
which then takes care of the rest.

The hybrid_fem usecase had a different issue, since it was making an inclusion
of a header_file dependent on whether __CUDACC__ is defined. With nvcc doing
the compilation both of the .cpp and .cu file in the usecase, the header got
included in two different compilation units, and functions were duplicated.
The issue was fixed by using a special define as guard which is set at the
top of the .cu file.

Going forward we recommend not to use special .cu files but compile everything
with nvcc (e.g. using the nvcc_wrapper).